refactor(auditlog): rename event handlers to action type handlers
- Rename `*EventHandler` classes and files to `*ActionTypeHandler`.
- Update base class `GuildAuditLogEntryCreateEventHandler` to `GuildAuditLogEntryCreateEventActionTypeHandler`.
- Adjust method names and Javadoc to align with `ActionType` handling.
