+++ src/main/java/io/github/eggy03/papertrail/bot/handlers/auditlog/AutoModActionTypeHandler.java
@@ -1,6 +1,6 @@
 package io.github.eggy03.papertrail.bot.handlers.auditlog;
 
-import io.github.eggy03.papertrail.bot.listeners.auditlog.GuildAuditLogEntryCreateEventHandler;
+import io.github.eggy03.papertrail.bot.listeners.auditlog.GuildAuditLogEntryCreateEventActionTypeHandler;
 import io.github.eggy03.papertrail.bot.utils.BooleanUtils;
 import io.github.eggy03.papertrail.bot.utils.auditlog.AutoModUtils;
 import io.github.eggy03.papertrail.sdk.client.AuditLogRegistrationClient;
@@ -23,12 +23,12 @@ import java.awt.Color;
 @ApplicationScoped
 @Slf4j
 @SuppressWarnings("java:S1192")
-public final class AutoModEventHandler extends GuildAuditLogEntryCreateEventHandler {
+public final class AutoModActionTypeHandler extends GuildAuditLogEntryCreateEventActionTypeHandler {
 
     private final @NonNull AuditLogRegistrationClient client;
 
     @Inject
-    public AutoModEventHandler(@NonNull AuditLogRegistrationClient client) {
+    public AutoModActionTypeHandler(@NonNull AuditLogRegistrationClient client) {
         this.client = client;
     }
