+++ src/main/java/io/github/eggy03/papertrail/bot/handlers/auditlog/ChannelActionTypeHandler.java
@@ -1,6 +1,6 @@
 package io.github.eggy03.papertrail.bot.handlers.auditlog;
 
-import io.github.eggy03.papertrail.bot.listeners.auditlog.GuildAuditLogEntryCreateEventHandler;
+import io.github.eggy03.papertrail.bot.listeners.auditlog.GuildAuditLogEntryCreateEventActionTypeHandler;
 import io.github.eggy03.papertrail.bot.utils.BooleanUtils;
 import io.github.eggy03.papertrail.bot.utils.DurationUtils;
 import io.github.eggy03.papertrail.bot.utils.auditlog.ChannelUtils;
@@ -24,12 +24,12 @@ import java.awt.Color;
 @ApplicationScoped
 @Slf4j
 @SuppressWarnings("java:S1192")
-public final class ChannelEventHandler extends GuildAuditLogEntryCreateEventHandler {
+public final class ChannelActionTypeHandler extends GuildAuditLogEntryCreateEventActionTypeHandler {
 
     private final @NonNull AuditLogRegistrationClient client;
 
     @Inject
-    public ChannelEventHandler(@NonNull AuditLogRegistrationClient client) {
+    public ChannelActionTypeHandler(@NonNull AuditLogRegistrationClient client) {
         this.client = client;
     }
