+++ src/main/java/io/github/eggy03/papertrail/bot/handlers/auditlog/EmojiActionTypeHandler.java
@@ -1,6 +1,6 @@
 package io.github.eggy03.papertrail.bot.handlers.auditlog;
 
-import io.github.eggy03.papertrail.bot.listeners.auditlog.GuildAuditLogEntryCreateEventHandler;
+import io.github.eggy03.papertrail.bot.listeners.auditlog.GuildAuditLogEntryCreateEventActionTypeHandler;
 import io.github.eggy03.papertrail.sdk.client.AuditLogRegistrationClient;
 import io.github.eggy03.papertrail.sdk.entity.AuditLogRegistrationEntity;
 import jakarta.enterprise.context.ApplicationScoped;
@@ -20,12 +20,12 @@ import java.awt.Color;
 @ApplicationScoped
 @Slf4j
 @SuppressWarnings("java:S1192")
-public final class IntegrationEventHandler extends GuildAuditLogEntryCreateEventHandler {
+public final class EmojiActionTypeHandler extends GuildAuditLogEntryCreateEventActionTypeHandler {
 
     private final @NonNull AuditLogRegistrationClient client;
 
     @Inject
-    public IntegrationEventHandler(@NonNull AuditLogRegistrationClient client) {
+    public EmojiActionTypeHandler(@NonNull AuditLogRegistrationClient client) {
         this.client = client;
     }
 
@@ -49,7 +49,7 @@ public final class IntegrationEventHandler extends GuildAuditLogEntryCreateEvent
     }
 
     @Override
-    public void onIntegrationCreate(@NonNull GuildAuditLogEntryCreateEvent event) {
+    public void onEmojiCreate(@NonNull GuildAuditLogEntryCreateEvent event) {
         String channelIdToSendTo = getRegisteredGuildChannel(event.getGuild().getId());
         if (channelIdToSendTo.isBlank()) return;
 
@@ -59,23 +59,21 @@ public final class IntegrationEventHandler extends GuildAuditLogEntryCreateEvent
         String mentionableExecutor = (executor != null ? executor.getAsMention() : ale.getUserId());
 
         EmbedBuilder eb = new EmbedBuilder();
-        eb.setTitle("Audit Log Entry | Integration Create Event");
-        eb.setDescription(MarkdownUtil.quoteBlock("Integration Created By: " + mentionableExecutor));
-        eb.setColor(Color.PINK);
+        eb.setTitle("Audit Log Entry | Emoji Create Event");
+        eb.setDescription(MarkdownUtil.quoteBlock("Emoji Created By: " + mentionableExecutor));
+        eb.setColor(Color.GREEN);
 
         ale.getChanges().forEach((changeKey, changeValue) -> {
 
             Object oldValue = changeValue.getOldValue();
             Object newValue = changeValue.getNewValue();
 
-            switch (changeKey) {
-                case "type" -> eb.addField(MarkdownUtil.underline("Integration Type"), "╰┈➤" + newValue, false);
-                case "name" -> eb.addField(MarkdownUtil.underline("Integration Name"), "╰┈➤" + newValue, false);
-
-                default -> {
-                    eb.addField("Unimplemented Change Key", changeKey, false);
-                    log.info("Unimplemented Change Key on Integration Create: {}\nOLD_VALUE: {}\nNEW_VALUE: {}", changeKey, oldValue, newValue);
-                }
+            if (changeKey.equals("name")) {
+                eb.addField(MarkdownUtil.underline("Emoji Name"), "╰┈➤" + newValue, false);
+                eb.addField(MarkdownUtil.underline("Emoji"), "╰┈➤" + "<:" + newValue + ":" + ale.getTargetId() + ">", false); // ale's TargetID retrieves the ID of the created emoji
+            } else {
+                eb.addField("Unimplemented Change Key", changeKey, false);
+                log.info("Unimplemented Change Key on Emoji Create: {}\nOLD_VALUE: {}\nNEW_VALUE: {}", changeKey, oldValue, newValue);
             }
         });
 
@@ -86,32 +84,7 @@ public final class IntegrationEventHandler extends GuildAuditLogEntryCreateEvent
     }
 
     @Override
-    public void onIntegrationUpdate(@NonNull GuildAuditLogEntryCreateEvent event) {
-
-        log.warn("Integration Update Event Detected. Implement this sometime later\n{}", event.getEntry().getChanges());
-
-        String channelIdToSendTo = getRegisteredGuildChannel(event.getGuild().getId());
-        if (channelIdToSendTo.isBlank()) return;
-
-        AuditLogEntry ale = event.getEntry();
-
-        EmbedBuilder eb = new EmbedBuilder();
-        eb.setTitle("Audit Log Entry | Integration Update Event");
-        eb.setColor(Color.LIGHT_GRAY);
-
-        String implementationNotice = "We do not have sufficient data to log the changes in an INTEGRATION_UPDATE Event."
-                .concat(" A proper implementation might happen in future releases if such an event is fired consistently.");
-
-        eb.addField("Implementation Notice", MarkdownUtil.codeblock(implementationNotice), false);
-
-        eb.setFooter("Audit Log Entry ID: " + ale.getId());
-        eb.setTimestamp(ale.getTimeCreated());
-
-        performChecksThenBuildAndSendEmbed(event, eb, channelIdToSendTo);
-    }
-
-    @Override
-    public void onIntegrationDelete(@NonNull GuildAuditLogEntryCreateEvent event) {
+    public void onEmojiUpdate(@NonNull GuildAuditLogEntryCreateEvent event) {
         String channelIdToSendTo = getRegisteredGuildChannel(event.getGuild().getId());
         if (channelIdToSendTo.isBlank()) return;
 
@@ -120,23 +93,23 @@ public final class IntegrationEventHandler extends GuildAuditLogEntryCreateEvent
         User executor = ale.getJDA().getUserById(ale.getUserIdLong());
         String mentionableExecutor = (executor != null ? executor.getAsMention() : ale.getUserId());
 
+
         EmbedBuilder eb = new EmbedBuilder();
-        eb.setTitle("Audit Log Entry | Integration Delete Event");
-        eb.setDescription(MarkdownUtil.quoteBlock("Integration Deleted By: " + mentionableExecutor));
-        eb.setColor(Color.MAGENTA);
+        eb.setTitle("Audit Log Entry | Emoji Update Event");
+        eb.setDescription(MarkdownUtil.quoteBlock("Emoji Updated By: " + mentionableExecutor));
+        eb.setColor(Color.YELLOW);
 
         ale.getChanges().forEach((changeKey, changeValue) -> {
+
             Object oldValue = changeValue.getOldValue();
             Object newValue = changeValue.getNewValue();
 
-            switch (changeKey) {
-                case "type" -> eb.addField(MarkdownUtil.underline("Integration Type"), "╰┈➤" + oldValue, false);
-                case "name" -> eb.addField(MarkdownUtil.underline("Integration Name"), "╰┈➤" + oldValue, false);
-
-                default -> {
-                    eb.addField("Unimplemented Change Key", changeKey, false);
-                    log.info("Unimplemented Change Key on Integration Delete: {}\nOLD_VALUE: {}\nNEW_VALUE: {}", changeKey, oldValue, newValue);
-                }
+            if (changeKey.equals("name")) {
+                eb.addField(MarkdownUtil.underline("Emoji Name Updated"), "╰┈➤" + "From " + oldValue + " to " + newValue, false);
+                eb.addField(MarkdownUtil.underline("Target Emoji"), "╰┈➤" + "<:" + newValue + ":" + ale.getTargetId() + ">", false);
+            } else {
+                eb.addField("Unimplemented Change Key", changeKey, false);
+                log.info("Unimplemented Change Key on Emoji Update: {}\nOLD_VALUE: {}\nNEW_VALUE: {}", changeKey, oldValue, newValue);
             }
         });
 
@@ -147,7 +120,7 @@ public final class IntegrationEventHandler extends GuildAuditLogEntryCreateEvent
     }
 
     @Override
-    public void onApplicationCommandPrivilegesUpdate(@NonNull GuildAuditLogEntryCreateEvent event) {
+    public void onEmojiDelete(@NonNull GuildAuditLogEntryCreateEvent event) {
         String channelIdToSendTo = getRegisteredGuildChannel(event.getGuild().getId());
         if (channelIdToSendTo.isBlank()) return;
 
@@ -157,11 +130,23 @@ public final class IntegrationEventHandler extends GuildAuditLogEntryCreateEvent
         String mentionableExecutor = (executor != null ? executor.getAsMention() : ale.getUserId());
 
         EmbedBuilder eb = new EmbedBuilder();
-        eb.setTitle("Audit Log Entry | Application Command Privilege Update");
-        eb.setDescription(MarkdownUtil.quoteBlock("Application Command Privilege Updated By: " + mentionableExecutor));
-        eb.setColor(Color.PINK);
+        eb.setTitle("Audit Log Entry | Emoji Delete Event");
+        eb.setDescription(MarkdownUtil.quoteBlock("Emoji Deleted By: " + mentionableExecutor));
+        eb.setColor(Color.RED);
 
-        eb.addField(MarkdownUtil.underline("More Info"), MarkdownUtil.codeblock("To know more about what changes were made, visit the Integrations section"), false);
+        ale.getChanges().forEach((changeKey, changeValue) -> {
+
+            Object oldValue = changeValue.getOldValue();
+            Object newValue = changeValue.getNewValue();
+
+            if (changeKey.equals("name")) {
+                eb.addField(MarkdownUtil.underline("Emoji Name"), "╰┈➤" + oldValue, false);
+            } else {
+                eb.addField("Unimplemented Change Key", changeKey, false);
+                log.info("Unimplemented Change Key on Emoji Delete: {}\nOLD_VALUE: {}\nNEW_VALUE: {}", changeKey, oldValue, newValue);
+            }
+        });
+        eb.addField("Deleted Emoji ID", "╰┈➤" + ale.getTargetId(), false);
 
         eb.setFooter("Audit Log Entry ID: " + ale.getId());
         eb.setTimestamp(ale.getTimeCreated());
