+++ src/main/java/io/github/eggy03/papertrail/bot/handlers/auditlog/ThreadActionTypeHandler.java
@@ -1,6 +1,6 @@
 package io.github.eggy03.papertrail.bot.handlers.auditlog;
 
-import io.github.eggy03.papertrail.bot.listeners.auditlog.GuildAuditLogEntryCreateEventHandler;
+import io.github.eggy03.papertrail.bot.listeners.auditlog.GuildAuditLogEntryCreateEventActionTypeHandler;
 import io.github.eggy03.papertrail.bot.utils.BooleanUtils;
 import io.github.eggy03.papertrail.bot.utils.DurationUtils;
 import io.github.eggy03.papertrail.bot.utils.auditlog.ChannelUtils;
@@ -25,12 +25,12 @@ import java.awt.Color;
 @ApplicationScoped
 @Slf4j
 @SuppressWarnings("java:S1192")
-public final class ThreadEventHandler extends GuildAuditLogEntryCreateEventHandler {
+public final class ThreadActionTypeHandler extends GuildAuditLogEntryCreateEventActionTypeHandler {
 
     private final @NonNull AuditLogRegistrationClient client;
 
     @Inject
-    public ThreadEventHandler(@NonNull AuditLogRegistrationClient client) {
+    public ThreadActionTypeHandler(@NonNull AuditLogRegistrationClient client) {
         this.client = client;
     }
