+++ src/main/java/io/github/eggy03/papertrail/bot/handlers/auditlog/UnimplementedActionTypeHandler.java
@@ -1,6 +1,6 @@
 package io.github.eggy03.papertrail.bot.handlers.auditlog;
 
-import io.github.eggy03.papertrail.bot.listeners.auditlog.GuildAuditLogEntryCreateEventHandler;
+import io.github.eggy03.papertrail.bot.listeners.auditlog.GuildAuditLogEntryCreateEventActionTypeHandler;
 import jakarta.enterprise.context.ApplicationScoped;
 import lombok.NonNull;
 import lombok.extern.slf4j.Slf4j;
@@ -10,10 +10,10 @@ import net.dv8tion.jda.api.events.guild.GuildAuditLogEntryCreateEvent;
 @ApplicationScoped
 @Slf4j
 @SuppressWarnings("java:S1192")
-public final class UnimplementedEventHandler extends GuildAuditLogEntryCreateEventHandler {
+public final class UnimplementedActionTypeHandler extends GuildAuditLogEntryCreateEventActionTypeHandler {
 
     @Override
-    public void onUnimplementedEvent(@NonNull GuildAuditLogEntryCreateEvent event) {
+    public void onUnimplementedActionType(@NonNull GuildAuditLogEntryCreateEvent event) {
         log.warn("Unimplemented Action Type: {} with Target Type: {}", event.getEntry().getType(), event.getEntry().getType().getTargetType());
     }
 }
